Update browser UI debugging
I've attempted to update this section as it seemed to be quite out of date. We have awesome builtin tools now. I've never used Jade in my life and can only hope my attempt at a link inside a li is correct. The internet said I'm not crazy and it's somehow non-trivial ( http://stackoverflow.com/questions/6989653/jade-links-inside-a-paragraph )